move formatCorrency util to currency-utils
