+++ frontend/app/utils/string-utils.ts
@@ -56,14 +56,3 @@ export function padWithZero(value: number, maxLength: number): string {
 export function trimToUndefined(str: string | undefined): string | undefined {
   return Number(str?.trim().length) > 0 ? str : undefined;
 }
-
-/**
- * Returns a formatted canadian currency string base on the
- * passed language's corresponding locale.
- */
-export function formatCurrency(number: number, lang: Language): string {
-  return number.toLocaleString(lang === 'fr' ? 'fr-CA' : 'en-CA', {
-    style: 'currency',
-    currency: 'CAD',
-  });
-}
